widen ints to 64bit

+++ src/Interpreter/Main.cpp
@@ -40,7 +40,7 @@ int main(int argc, char** argv) {
 // https://github.com/brwhale/KataScript/blob/main/README.md#c-integration
 //-------------------------------------------------------------------------------------------------//
 
-int integrationExample(int a, int b) {	
+int64_t integrationExample(int64_t a, int64_t b) {
 	return (a * b) + b;
 }
 
@@ -72,11 +72,11 @@ void integrationExample() {
 	auto varRef = interp.resolveVariable("i");
 
 	// or just call a function directly
-	varRef = interp.callFunction(newfunc, 4, 3);
+	varRef = interp.callFunction(newfunc, 4ll, 3ll);
 
 	// Setp 3: Unwrap your result
 	// if the type is known
-	int i = varRef->getInt();
+    int64_t i = varRef->getInt();
     std::cout << i << "\n";
 
 	// visit style
